feat(onboarding): split git connect and repo select into separate steps
Splits the combined "pick repo + connect GitHub" step into two pages:

- `connect-git` (optional) — merges the existing CLI install check page
  with the GitHub OAuth panel. Explains what git integration unlocks
  (cloud sandboxes, branch pushes, in-app PR review).
- `select-repo` (final) — folder picker only, with an explicit
  "Skip & get started" path. When skipped, `useSetupDiscovery` already
  no-ops on empty `selectedDirectory`, so enricher-based first-task
  suggestions naturally skip too.

Also retires the `tools_not_installed` skip reason in favor of
`no_repo_selected`, and renames `cli_skipped` → `repo_skipped` on the
ONBOARDING_COMPLETED event to match the new final step.

Generated-By: PostHog Code
Task-Id: d15be8ed-c5d9-44cc-8cc7-de90e1e84f2e
