+++ apps/code/src/renderer/features/onboarding/components/OnboardingFlow.tsx
@@ -14,10 +14,10 @@ import { useEffect, useRef } from "react";
 import { useHotkeys } from "react-hotkeys-hook";
 
 import { useOnboardingFlow } from "../hooks/useOnboardingFlow";
-import { CliInstallStep } from "./CliInstallStep";
-import { GitIntegrationStep } from "./GitIntegrationStep";
+import { ConnectGitStep } from "./ConnectGitStep";
 import { InviteCodeStep } from "./InviteCodeStep";
 import { ProjectSelectStep } from "./ProjectSelectStep";
+import { SelectRepoStep } from "./SelectRepoStep";
 import { StepIndicator } from "./StepIndicator";
 import { WelcomeScreen } from "./WelcomeScreen";
 
@@ -115,12 +115,12 @@ export function OnboardingFlow() {
   useHotkeys("right", handleNext, { enableOnFormTags: false }, [handleNext]);
   useHotkeys("left", handleBack, { enableOnFormTags: false }, [handleBack]);
 
-  const handleComplete = (cliSkipped: boolean) => {
-    if (cliSkipped) {
+  const handleComplete = (repoSkipped: boolean) => {
+    if (repoSkipped) {
       track(ANALYTICS_EVENTS.ONBOARDING_STEP_SKIPPED, {
         step_id: currentStep,
         step_index: currentIndex,
-        reason: "tools_not_installed",
+        reason: "no_repo_selected",
       });
     } else {
       trackStepCompleted();
@@ -130,7 +130,7 @@ export function OnboardingFlow() {
         (Date.now() - flowStartedAtRef.current) / 1000,
       ),
       github_connected: githubUserIntegrations.length > 0,
-      cli_skipped: cliSkipped,
+      repo_skipped: repoSkipped,
     });
     completeOnboarding();
     navigateToTaskInput();
@@ -235,9 +235,9 @@ export function OnboardingFlow() {
             </motion.div>
           )}
 
-          {currentStep === "github" && (
+          {currentStep === "connect-git" && (
             <motion.div
-              key="github"
+              key="connect-git"
               custom={direction}
               initial="enter"
               animate="center"
@@ -246,29 +246,29 @@ export function OnboardingFlow() {
               transition={{ duration: 0.3 }}
               className="min-h-0 w-full flex-1"
             >
-              <GitIntegrationStep
-                onNext={handleNext}
+              <ConnectGitStep onNext={handleNext} onBack={handleBack} />
+            </motion.div>
+          )}
+
+          {currentStep === "select-repo" && (
+            <motion.div
+              key="select-repo"
+              custom={direction}
+              initial="enter"
+              animate="center"
+              exit="exit"
+              variants={stepVariants}
+              transition={{ duration: 0.3 }}
+              className="min-h-0 w-full flex-1"
+            >
+              <SelectRepoStep
+                onComplete={handleComplete}
                 onBack={handleBack}
                 selectedDirectory={selectedDirectory}
                 detectedRepo={detectedRepo}
                 isDetectingRepo={isDetectingRepo}
                 onDirectoryChange={handleDirectoryChange}
               />
-            </motion.div>
-          )}
-
-          {currentStep === "install-cli" && (
-            <motion.div
-              key="install-cli"
-              custom={direction}
-              initial="enter"
-              animate="center"
-              exit="exit"
-              variants={stepVariants}
-              transition={{ duration: 0.3 }}
-              className="min-h-0 w-full flex-1"
-            >
-              <CliInstallStep onComplete={handleComplete} onBack={handleBack} />
             </motion.div>
           )}
         </AnimatePresence>
